requirements.txt
flask==3.1.0
qrcode==8.0
Pillow==11.1.0
openpyxl==3.1.5
